@@ -1,6 +1,3 @@
-[bdist_wheel]
-universal = 1
-
 [metadata]
 name = pyversion-info
 version = attr:pyversion_info.__version__
@@ -21,10 +18,8 @@ keywords =
 classifiers =
     Development Status :: 4 - Beta
     #Development Status :: 5 - Production/Stable
-    Programming Language :: Python :: 2
-    Programming Language :: Python :: 2.7
+    Programming Language :: Python :: 3 :: Only
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
@@ -44,8 +39,8 @@ project_urls =
 py_modules = pyversion_info
 package_dir =
     =src
-python_requires = >=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*, <4
+python_requires = ~=3.6
 install_requires =
     appdirs                 ~= 1.4
     cachecontrol[filecache] ~= 0.12.0
-    requests                == 2.*
+    requests                ~= 2.20
